Fixed identity key error when saving data into database from background worker

+++ APIAPP/BackgroundWorker/Models/PointsOfInterestDTO.cs
@@ -3,12 +3,12 @@
     public class PointsOfInterestDTO
     {
         public int Id { get; set; }
-        public int Radius { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public double Radius { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public bool Archived { get; set; }
 
-        public PointsOfInterestDTO(int id, int radius, int latitude, int longitude, bool archived = false) {
+        public PointsOfInterestDTO(int id, double radius, double latitude, double longitude, bool archived = false) {
             Id = id;
             Radius = radius; 
             Latitude = latitude;
